Refactor, tests, functional approach

+++ Services/Forum/Forum.Core/Entities/Answers/Commands/Delete/DeleteAnswerCommand.cs
@@ -1,10 +1,12 @@
-﻿using Forum.Core.Services;
+﻿using Forum.Core.Results;
+using Forum.Core.Services;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
+using OneOf;
 
 namespace Forum.Core.Entities.Answers.Commands.Delete;
 
-public class DeleteAnswerCommand : IRequest
+public class DeleteAnswerCommand : IRequest<OneOf<Unit, NotFoundResult, InvalidDomainBehaviorResult>>
 {
     public long Id { get; }
 
@@ -14,7 +16,7 @@ public class DeleteAnswerCommand : IRequest
     }
 }
 
-public class DeleteAnswerCommandHandler : IRequestHandler<DeleteAnswerCommand>
+public class DeleteAnswerCommandHandler : IRequestHandler<DeleteAnswerCommand, OneOf<Unit, NotFoundResult, InvalidDomainBehaviorResult>>
 {
     private readonly IDomainContext _context;
     private readonly IUserService _userService;
@@ -25,16 +27,19 @@ public class DeleteAnswerCommandHandler : IRequestHandler<DeleteAnswerCommand>
         _userService = userService;
     }
 
-    public async Task<Unit> Handle(DeleteAnswerCommand request, CancellationToken cancellationToken)
+    public async Task<OneOf<Unit, NotFoundResult, InvalidDomainBehaviorResult>> Handle(DeleteAnswerCommand request, CancellationToken cancellationToken)
     {
         var answer = await _context.Answers
             .FirstOrDefaultAsync(a => a.Id == request.Id, cancellationToken);
+        if (answer == null)
+            return new NotFoundResult($"There is no answer with such id: {request.Id}");
         
-        ThrowIf.Null(answer, $"There is no answer with such id: {request.Id}");
-        ThrowIf.Null(answer!.AuthorId, "Answer can't have no author");
+        if (string.IsNullOrEmpty(answer.AuthorId) || answer.Author == null)
+            return new InvalidDomainBehaviorResult("Answer must have an author");
 
         var whetherUserIsAuthor = answer.AuthorId == _userService.UserId;
-        ThrowIf.False(whetherUserIsAuthor, "You can't delete other user' answer");
+        if (!whetherUserIsAuthor)
+            return new InvalidDomainBehaviorResult("You can't delete other user' answer");
 
         _context.Answers.Remove(answer);
         await _context.SaveChangesAsync(cancellationToken);
